requirements.txt
frida==16.0.8
frida-tools==12.0.4
click==8.1.3
r2pipe==1.7.4
requests==2.28.2
androguard==3.3.5
pyjson5==1.6.2
pyyaml==6.0
intervaltree==3.1.0
pyre2==0.3.6
google-play-scraper==1.2.3
tidevice==0.10.7
psutil==5.9.5
Flask==2.3.2
xmltodict==0.13.0
pyshark==0.6
pandas==2.0.1
pandasql==0.7.3
paramiko==3.2.0
scp==0.14.5
tqdm==4.65.0